feat: Implement bash-style grid completion
completion style can be toggled using `shopt prompt.complete_style`

valid options for `prompt.complete_style` are 'grid' or 'fuzzy'

added prompt.complete_style to the nix module options

made grid completion the default completion style
