+++ doc/builtin.txt
@@ -1445,6 +1445,7 @@ SHOPT                                            *builtin-shopt* *shopt*
     `screensaver_cmd`         command to run as screensaver
     `screensaver_idle_time`   idle seconds before screensaver (0=off)
     `trunc_prompt_path`       max path segments in `\W` prompt escape
+    `complete_style`          completion menu style (`grid` or `fuzzy`)
 
   Set options (`set.*`) ~
